11-intermediate-scripting/cli-app/requirements.txt
python-dotenv==0.15.0
psutil==5.8.0
pymongo==3.11.3
pymongo[srv]
dnspython

